feat(axon_rust-pkl.10): add pending-job caps for embed, extract, and ingest queues
Generalize check_pending_cap() into check_pending_cap_for() and add per-queue
wrappers for embed (AXON_MAX_PENDING_EMBED_JOBS, default 50), extract
(AXON_MAX_PENDING_EXTRACT_JOBS, default 50), and ingest
(AXON_MAX_PENDING_INGEST_JOBS, default 50). Setting any cap to 0 disables it.
Error messages identify the queue name and env var. Cap tests injected via
parameter (no unsafe env mutation). Tests extracted to ops/tests.rs per
Rust 2018 file-per-module layout.
