+++ .env.example
@@ -153,6 +153,12 @@ AXON_EMBED_DOC_TIMEOUT_SECS=300
 
 # [OPTIONAL] Max pending crawl jobs before new submissions are rejected (0 = unlimited). Default: 100.
 AXON_MAX_PENDING_CRAWL_JOBS=100
+# [OPTIONAL] Max pending embed jobs before new submissions are rejected (0 = unlimited). Default: 50.
+AXON_MAX_PENDING_EMBED_JOBS=50
+# [OPTIONAL] Max pending extract jobs before new submissions are rejected (0 = unlimited). Default: 50.
+AXON_MAX_PENDING_EXTRACT_JOBS=50
+# [OPTIONAL] Max pending ingest jobs before new submissions are rejected (0 = unlimited). Default: 50.
+AXON_MAX_PENDING_INGEST_JOBS=50
 
 # [OPTIONAL] Seconds before a running job is considered stale. Default: 300.
 AXON_JOB_STALE_TIMEOUT_SECS=300
@@ -213,8 +219,11 @@ AXON_ACP_AUTO_APPROVE=true
 # [OPTIONAL] Prewarm ACP adapter on startup to eliminate cold-start latency (~45s). Default: true.
 AXON_ACP_PREWARM=true
 
-# [OPTIONAL] Max concurrent ACP completion sessions. Default: none (unlimited).
+# [OPTIONAL] Max concurrent ACP completion requests. Default: none (unlimited).
 AXON_ACP_COMPLETION_CONCURRENCY=
+
+# [OPTIONAL] Max cached ACP sessions (guards memory; distinct from concurrency). Default: 100 (0 = unlimited).
+AXON_ACP_MAX_SESSIONS=100
 
 # [OPTIONAL] Remote ACP: route completions through a remote axon serve WebSocket.
 # Set to the base HTTP/HTTPS URL of the remote instance.
